+++ sdk/keyvault/azure-security-keyvault-certificates/customizations/src/main/java/CertificatesCustomizations.java
@@ -104,7 +104,7 @@ public class CertificatesCustomizations extends Customization {
             .setJavadocComment("The versions of Azure Key Vault Certificates supported by this client library.");
 
         for (String version : Arrays.asList("7.0", "7.1", "7.2", "7.3", "7.4", "7.5", "7.6",
-            "2025-07-01")) {
+            "2025-07-01", "2026-03-01-preview")) {
             enumDeclaration.addEnumConstant("V" + version.replace('.', '_').replace('-', '_').toUpperCase())
                 .setJavadocComment("Service version {@code " + version + "}.")
                 .addArgument(new StringLiteralExpr(version));
@@ -125,7 +125,7 @@ public class CertificatesCustomizations extends Customization {
             .setType("CertificateServiceVersion")
             .setJavadocComment(new Javadoc(parseText("Gets the latest service version supported by this client library."))
                 .addBlockTag("return", "The latest {@link CertificateServiceVersion}."))
-            .setBody(StaticJavaParser.parseBlock("{ return V2025_07_01; }"));
+            .setBody(StaticJavaParser.parseBlock("{ return V2026_03_01_PREVIEW; }"));
 
         customization.getRawEditor()
             .addFile("src/main/java/com/azure/security/keyvault/certificates/CertificateServiceVersion.java",